+++ src/main/java/za/ac/cput/repository/business/CartRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import za.ac.cput.domain.business.Cart;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
 
 
 }
